feat: add centralised comparators
