+++ app/server/appsmith-server/src/main/java/com/appsmith/server/helpers/AppsmithComparators.java
@@ -1,0 +1,6 @@
+package com.appsmith.server.helpers;
+
+import com.appsmith.server.helpers.ce.AppsmithComparatorsCE;
+
+public class AppsmithComparators extends AppsmithComparatorsCE {
+}
